requirements.txt
bs4
undetected-chromedriver
Pillow
selenium
easyocr
